--- a/Applications/shapeworks/shapeworks.cpp
+++ b/Applications/shapeworks/shapeworks.cpp
@@ -5,6 +5,7 @@
 #include <exception>
 #include "Executable.h"
 #include "Commands.h"
+#include <Profiling.h>
 
 using namespace shapeworks;
 
@@ -111,9 +112,14 @@ int main(int argc, char *argv[])
   shapeworks.addCommand(ConvertProjectCommand::getCommand());
 
   try {
-    return shapeworks.run(argc, argv);
+    TIME_START("shapeworks");
+    int rc = shapeworks.run(argc, argv);
+    TIME_STOP("shapeworks");
+    TIME_FINALIZE();
+    return rc;
   } catch (const std::exception &e) {
     std::cout << e.what() << std::endl;
+    TIME_FINALIZE();
     return EXIT_FAILURE;
   }
 }