… Libs/Common/CMakeLists.txt
… Libs/Optimize/GradientDescentOptimizer.cpp
… Libs/Optimize/Optimize.cpp
… Libs/Project/Project.cpp
… Libs/Utils/CMakeLists.txt
… Libs/Utils/PlatformUtils.h
… Python/shapeworks/shapeworks/shape_scalars.py
… Studio/Analysis/AnalysisTool.cpp
… Studio/Data/Session.cpp
… Studio/Job/ShapeScalarJob.cpp
… Studio/Job/ShapeScalarJob.h